app-ocr/requirements.txt
flask
pytesseract
pillow